pyproject.toml
[project]
name = "cbig"
version = "1.3.0"
description = "CodeBase Insight Generator - Cross-language CLI extractor for directory-level code analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-java>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-html>=0.20.0",
    "tree-sitter-rust>=0.21.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.8.0",
    "pathspec>=0.12.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0"
]

[project.scripts]
cbig = "cbig.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cbig"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md"
]
